Added an undo button to the visual scorekeeping view.
The undo button removes the last placed disc for the active player.

This is slightly awkward. Consider when discs have been created for player one, then player two, then player one again. Tapping the undo button twice removes the two player one discs, rather than the two latest discs, which is the expected behaviour.

Also changed the disc collections to mutable arrays rather than sets, so that order is maintained. The initializing of these disc collections should be DRYer. Tracking down all the places in the code where the initialization implementation needed to be changed was a pain.