Changed discThatCollidesWithDiscAtPosition to return a NSValue pointer so that nil can be returned rather than a point at -(1, -1).
I considered changing adjustedPositionForNewDisc to also return a NSValue pointer, but decided against it to avoid the added boilerplate in the calling code.

The general rule should be: return a NSValue pointer from methods that can return an invalid point and use nil to signify that invalid point.

For methods that only return valid points, just return a CGPoint directly.